+++ backend/middlewares/auth.ts
@@ -1,20 +1,11 @@
 import { Request } from "express";
 import { AuthenticationError, ExpressContext } from "apollo-server-express";
 import { GraphQLResolveInfo } from "graphql";
-import { IMiddlewareFunction } from "graphql-middleware";
 
 import AuthService from "../services/implementations/authService";
 import UserService from "../services/implementations/userService";
 import IAuthService from "../services/interfaces/authService";
 import { Role } from "../types";
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-type Resolve = (
-  parent: any,
-  args: { [key: string]: any },
-  context: ExpressContext | undefined,
-  info: GraphQLResolveInfo | undefined,
-) => any;
 
 const authService: IAuthService = new AuthService(new UserService());
 
@@ -30,51 +21,12 @@ export const getAccessToken = (req: Request): string | null => {
   return null;
 };
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-const callResolver = async (
-  resolver: IMiddlewareFunction<any, ExpressContext>,
-  resolve: Resolve,
-  parent: any,
-  args: { [key: string]: any },
-  context: ExpressContext,
-  info: GraphQLResolveInfo,
-  // eslint-disable-next-line consistent-return
-) => {
-  if ("resolve" in resolver) {
-    return resolver.resolve?.(resolve, parent, args, context, info);
-  }
-  if (typeof resolver === "function") {
-    return resolver(resolve, parent, args, context, info);
-  }
-};
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-export const and = (
-  firstFn: IMiddlewareFunction<any, ExpressContext>,
-  secondFn: IMiddlewareFunction<any, ExpressContext>,
-) => {
-  return async (
-    resolve: Resolve,
-    parent: any,
-    args: { [key: string]: any },
-    context: ExpressContext,
-    info: GraphQLResolveInfo,
-  ) => {
-    const resolveFirst = async () => {
-      return callResolver(secondFn, resolve, parent, args, context, info);
-    };
-    return callResolver(firstFn, resolveFirst, parent, args, context, info);
-  };
-};
-
 /* Determine if the user whose role is to be changed does not match the currently 
 logged in user. 
  * Note: userIdField is the name of the request parameter containing the requested userId */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-export const isAuthorizedToChangeRole = (userIdField: string) => {
+export const iDNotSameasActiveUser = (userIdField: string) => {
   return async (
     resolve: (
       parent: any,
@@ -90,17 +42,11 @@ export const isAuthorizedToChangeRole = (userIdField: string) => {
     const accessToken = getAccessToken(context.req);
     const authorized =
       accessToken &&
-      (await authService.isAuthorizedToChangeRole(
-        accessToken,
-        args[userIdField],
-      ));
+      (await authService.iDNotSameasActiveUser(accessToken, args[userIdField]));
 
     if (!authorized) {
-      throw new AuthenticationError(
-        "Failed authentication and/or authorization by userId",
-      );
+      throw new AuthenticationError("User ID cannot change its own role");
     }
-
     return resolve(parent, args, context, info);
   };
 };
@@ -130,7 +76,6 @@ export const isAuthorizedByRole = (roles: Set<Role>) => {
         "Failed authentication and/or authorization by role",
       );
     }
-
     return resolve(parent, args, context, info);
   };
 };
